requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
curl-cffi>=0.6.0
